@@ -1,6 +1,6 @@
 {
     "name": "wekser/laragram",
-    "description": "Extend Laravel and Lumen interface for the quick and easy development of Telegram bots.",
+    "description": "Laravel framework for easy develop Telegram Bot.",
     "type": "library",
     "license": "MIT",
     "homepage": "https://github.com/wekser/laragram",
@@ -21,15 +21,18 @@
         }
     ],
     "require": {
-        "php": "^7.3|^8.0",
-        "guzzlehttp/guzzle": "^7.0.1"
+        "php": "^8.0",
+        "guzzlehttp/guzzle": "^7.2",
+        "illuminate/support": "^8.0|^9.0",
+        "nesbot/carbon": "^2.63"
     },
     "require-dev": {
-        "mockery/mockery": "^1.4.4",
-        "phpunit/phpunit": "^9.5.10"
+        "laravel/framework": "^9.0",
+        "mockery/mockery": "^1.4",
+        "phpunit/phpunit": "^9.5"
     },
     "config": {
-        "optimize-autoloader": true,
+        "sort-packages": true,
         "preferred-install": "dist"
     },
     "autoload": {
@@ -56,5 +59,7 @@
                 "BotResponse": "Wekser\\Laragram\\Facades\\BotResponse"
             }
         }
-    }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
