@@ -19,7 +19,15 @@ return [
 
     'auth' => [
         'driver' => 'database',
-        'model' => \Wekser\Laragram\Models\User::class
+        'session' => [
+            'lifetime' => 10080, // 7 days in minutes
+            'model' => \Wekser\Laragram\Models\Session::class,
+            'table' => 'laragram_sessions'
+        ],
+        'user' => [
+            'model' => \Wekser\Laragram\Models\User::class,
+            'table' => 'laragram_users'
+        ]
     ],
 
     /*
