@@ -20,11 +20,11 @@ use Wekser\Laragram\Facades\BotAuth;
 class BotResponse
 {
     /**
-     * The location to next user request.
+     * The station to next user request.
      *
      * @var string
      */
-    public $location;
+    public $station;
 
     /**
      * The contents of rendered view.
@@ -87,7 +87,7 @@ class BotResponse
      */
     public function user($user): self
     {
-        $model = config('laragram.auth.model');
+        $model = config('laragram.auth.user.model');
 
         $this->user = !($user instanceof $model) ?: $user;
 
@@ -95,14 +95,14 @@ class BotResponse
     }
 
     /**
-     * Set the location to next user request.
+     * Set the station to next user request.
      *
-     * @param string $location
+     * @param string $station
      * @return $this
      */
-    public function redirect(string $location): self
+    public function redirect(string $station): self
     {
-        $this->location = $location;
+        $this->station = $station;
 
         return $this;
     }
